# pkg/trace: remove custom aggregations