--- a/pkg/trace/agent/agent.go
+++ b/pkg/trace/agent/agent.go
@@ -69,7 +69,7 @@ func NewAgent(ctx context.Context, conf *config.AgentConfig) *Agent {
 
 	return &Agent{
 		Receiver:           api.NewHTTPReceiver(conf, dynConf, in),
-		Concentrator:       stats.NewConcentrator(conf.ExtraAggregators, conf.BucketInterval.Nanoseconds(), statsChan),
+		Concentrator:       stats.NewConcentrator(conf.BucketInterval.Nanoseconds(), statsChan),
 		Blacklister:        filters.NewBlacklister(conf.Ignore["resource"]),
 		Replacer:           filters.NewReplacer(conf.ReplaceTags),
 		ScoreSampler:       NewScoreSampler(conf),