--- a/pkg/trace/config/config.go
+++ b/pkg/trace/config/config.go
@@ -125,8 +125,7 @@ func New() *AgentConfig {
 		DefaultEnv: "none",
 		Endpoints:  []*Endpoint{{Host: "https://trace.agent.datadoghq.com"}},
 
-		BucketInterval:   time.Duration(10) * time.Second,
-		ExtraAggregators: []string{"http.status_code", "version", "_dd.hostname"},
+		BucketInterval: time.Duration(10) * time.Second,
 
 		ExtraSampleRate: 1.0,
 		MaxTPS:          10,