--- a/pkg/trace/test/testutil/stats.go
+++ b/pkg/trace/test/testutil/stats.go
@@ -18,7 +18,7 @@ const defaultEnv = "none"
 // TestBucket returns a fixed stats bucket to be used in unit tests
 func TestBucket() stats.Bucket {
 	srb := stats.NewRawBucket(0, 1e9)
-	srb.HandleSpan(TestWeightedSpan(), defaultEnv, defaultAggregators, nil)
+	srb.HandleSpan(TestWeightedSpan(), defaultEnv, nil)
 	sb := srb.Export()
 
 	// marshalling then unmarshalling data to:
@@ -42,7 +42,7 @@ func TestBucket() stats.Bucket {
 func BucketWithSpans(spans []*stats.WeightedSpan) stats.Bucket {
 	srb := stats.NewRawBucket(0, 1e9)
 	for _, s := range spans {
-		srb.HandleSpan(s, defaultEnv, defaultAggregators, nil)
+		srb.HandleSpan(s, defaultEnv, nil)
 	}
 	return srb.Export()
 }