… pkg/trace/stats/statsraw_test.go
… releasenotes/notes/trace-agent-remove-custom-aggregations-c36ddb504d4ff147.yaml