### What does this PR do?

Remove `apm_config.extra_aggregators` and switches to a struct to represent aggregations instead of a string.

It reduces the number of allocations, but increases the cost of hashing to update the maps. Which explains why there is no difference in execution time for the `HandleSpan` benchmark and a rather small one for `HandleSpanRandom`. The large difference visible for `HandleSpanSublayers` comes from the change introduced in commit e3ff9e0 which stores sublayers in the same map as other metrics.
 
```
name                   old time/op    new time/op    delta
HandleSpan-4             3.28µs ± 3%    3.28µs ± 4%      ~     (p=0.730 n=5+5)
HandleSpanSublayers-4    11.3µs ± 2%     4.8µs ± 8%   -57.84%  (p=0.008 n=5+5)
HandleSpanRandom-4       11.3µs ± 6%    10.2µs ± 3%    -9.32%  (p=0.008 n=5+5)

name                   old alloc/op   new alloc/op   delta
HandleSpan-4             1.15kB ± 0%    0.00kB       -100.00%  (p=0.008 n=5+5)
HandleSpanSublayers-4    3.70kB ± 0%    0.00kB       -100.00%  (p=0.008 n=5+5)
HandleSpanRandom-4       5.64kB ± 0%    1.23kB ± 0%   -78.29%  (p=0.008 n=5+5)

name                   old allocs/op  new allocs/op  delta
HandleSpan-4               16.0 ± 0%       0.0       -100.00%  (p=0.008 n=5+5)
HandleSpanSublayers-4      32.0 ± 0%       0.0       -100.00%  (p=0.008 n=5+5)
HandleSpanRandom-4         50.0 ± 0%      18.0 ± 0%   -64.00%  (p=0.008 n=5+5)
```


### Motivation

What inspired you to submit this pull request?

### Additional Notes

Anything else we should know when reviewing?

### Describe your test plan

Write there any instructions and details you may have to test your PR.